requirements.txt
Flask==3.0.3
gunicorn==21.2.0
pdf2image==1.17.0
PyPDF2==3.0.1
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.84
Flask-Cors==4.0.1
